+++ src/pdfkernel.cpp
@@ -3,7 +3,7 @@
 ** Purpose:     Implementation of wxPdfDocument (internal methods)
 ** Author:      Ulrich Telle
 ** Created:     2006-01-27
-** Copyright:   (c) 2006-2025 Ulrich Telle
+** Copyright:   (c) 2006-2026 Ulrich Telle
 ** Licence:     wxWindows licence
 ** SPDX-License-Identifier: LGPL-3.0+ WITH WxWindows-exception-3.1
 */
@@ -2358,139 +2358,296 @@ wxPdfDocument::PutPatterns()
   for (patternIter = m_patterns->begin(); patternIter != m_patterns->end(); patternIter++)
   {
     wxPdfPattern* pattern = patternIter->second;
-    NewObj(pattern->GetObjIndex());
-    Out("<<");
-    Out("/Type /Pattern");
-    Out("/PatternType 1");
-    Out("/PaintType 1");
-    Out("/TilingType 1");
-    if (pattern->GetPatternStyle() == wxPDF_PATTERNSTYLE_IMAGE ||
-        pattern->GetPatternStyle() == wxPDF_PATTERNSTYLE_TEMPLATE )
+    wxPdfPatternStyle patternStyle = pattern->GetPatternStyle();
+    switch(patternStyle)
     {
-      OutAscii(wxString(wxS("/BBox [0 0 ")) +
-        wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4) + wxS(" ") +
-        wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4) + wxS("]"));
-      OutAscii(wxString(wxS("/XStep ")) +
-        wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4));
-      OutAscii(wxString(wxS("/YStep ")) +
-        wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4));
-      if (pattern->GetPatternStyle() == wxPDF_PATTERNSTYLE_IMAGE)
-      {
-        wxPdfImage* image = pattern->GetImage();
-        OutAscii(wxString::Format(wxS("/Resources << /XObject << /I%d %d 0 R >> >>"), image->GetIndex(), image->GetObjIndex()));
-        Out("/Matrix [ 1 0 0 1 0 0 ]");
+      case wxPDF_PATTERNSTYLE_IMAGE:
+      case wxPDF_PATTERNSTYLE_TEMPLATE:
+        {
+          NewObj(pattern->GetObjIndex());
+          Out("<<");
+          Out("/Type /Pattern");
+          Out("/PatternType 1");
+          Out("/PaintType 1");
+          Out("/TilingType 1");
 
-        wxString sdata = wxString(wxS("q ")) +
-          wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4) + wxS(" 0 0 ") +
-          wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4) + wxS(" 0 0 cm ") +
-          wxString::Format(wxS("/I%d Do Q"), image->GetIndex());
-        wxMemoryOutputStream mos;
-        mos.Write(sdata.ToAscii(), sdata.Length());
-        OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
-        Out(">>");
-        PutStream(mos);
-      }
-      else
-      {
-        int templateId = pattern->GetTemplateId();
-        wxPdfTemplate* tpl = (*m_templates)[templateId];
-        OutAscii(wxString(wxS("/Resources << /XObject << ")) + m_templatePrefix + wxString::Format(wxS("%d %d 0 R >> >>"), tpl->GetIndex(), tpl->GetObjIndex()));
-        Out("/Matrix [ 1 0 0 1 0 0 ]");
-        double x, y, w, h;
-        w = pattern->GetWidth();
-        h = pattern->GetHeight();
-        x = tpl->GetX();
-        y = tpl->GetY();
-        GetTemplateSize(templateId, w, h);
+          OutAscii(wxString(wxS("/BBox [0 0 ")) +
+            wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4) + wxS(" ") +
+            wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4) + wxS("]"));
+          OutAscii(wxString(wxS("/XStep ")) +
+            wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4));
+          OutAscii(wxString(wxS("/YStep ")) +
+            wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4));
+          if (pattern->GetPatternStyle() == wxPDF_PATTERNSTYLE_IMAGE)
+          {
+            wxPdfImage* image = pattern->GetImage();
+            OutAscii(wxString::Format(wxS("/Resources << /XObject << /I%d %d 0 R >> >>"), image->GetIndex(), image->GetObjIndex()));
+            Out("/Matrix [ 1 0 0 1 0 0 ]");
 
-        double xScale = w / tpl->GetWidth();
-        double yScale = h / tpl->GetHeight();
-        double xTrans = (x - xScale * tpl->GetX()) * m_k;
-        double yTrans = (y - yScale * tpl->GetY()) * m_k;
-        wxString sdata = wxString(wxS("q ")) +
-          wxPdfUtility::Double2String(xScale, 4) + wxString(wxS(" 0 0 ")) +
-          wxPdfUtility::Double2String(yScale, 4) + wxString(wxS(" ")) +
-          wxPdfUtility::Double2String(xTrans, 2) + wxString(wxS(" ")) +
-          wxPdfUtility::Double2String(yTrans, 2) + wxString(wxS(" cm ")) +
-          m_templatePrefix + wxString::Format(wxS("%d Do Q"), tpl->GetIndex());
-        wxMemoryOutputStream mos;
-        mos.Write(sdata.ToAscii(), sdata.Length());
-        OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
-        Out(">>");
-        PutStream(mos);
-      }
+            wxString sdata = wxString(wxS("q ")) +
+              wxPdfUtility::Double2String(pattern->GetWidth() * m_k, 4) + wxS(" 0 0 ") +
+              wxPdfUtility::Double2String(pattern->GetHeight() * m_k, 4) + wxS(" 0 0 cm ") +
+              wxString::Format(wxS("/I%d Do Q"), image->GetIndex());
+            wxMemoryOutputStream mos;
+            mos.Write(sdata.ToAscii(), sdata.Length());
+            OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
+            Out(">>");
+            PutStream(mos);
+          }
+          else
+          {
+            int templateId = pattern->GetTemplateId();
+            wxPdfTemplate* tpl = (*m_templates)[templateId];
+            OutAscii(wxString(wxS("/Resources << /XObject << ")) + m_templatePrefix + wxString::Format(wxS("%d %d 0 R >> >>"), tpl->GetIndex(), tpl->GetObjIndex()));
+            Out("/Matrix [ 1 0 0 1 0 0 ]");
+            double x, y, w, h;
+            w = pattern->GetWidth();
+            h = pattern->GetHeight();
+            x = tpl->GetX();
+            y = tpl->GetY();
+            GetTemplateSize(templateId, w, h);
+
+            double xScale = w / tpl->GetWidth();
+            double yScale = h / tpl->GetHeight();
+            double xTrans = (x - xScale * tpl->GetX()) * m_k;
+            double yTrans = (y - yScale * tpl->GetY()) * m_k;
+            wxString sdata = wxString(wxS("q ")) +
+              wxPdfUtility::Double2String(xScale, 4) + wxString(wxS(" 0 0 ")) +
+              wxPdfUtility::Double2String(yScale, 4) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(xTrans, 2) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(yTrans, 2) + wxString(wxS(" cm ")) +
+              m_templatePrefix + wxString::Format(wxS("%d Do Q"), tpl->GetIndex());
+            wxMemoryOutputStream mos;
+            mos.Write(sdata.ToAscii(), sdata.Length());
+            OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
+            Out(">>");
+            PutStream(mos);
+          }
+          Out("endobj");
+        }
+        break;
+
+      case wxPDF_PATTERNSTYLE_LINEAR_GRADIENT:
+      case wxPDF_PATTERNSTYLE_RADIAL_GRADIENT:
+        {
+          wxGraphicsGradientStops stops = pattern->GetStops();
+          wxColour startColour = stops.GetStartColour();
+          wxColour stopColour;
+          size_t nStops = stops.GetCount() - 2;
+          int idFuncShading = m_n + 1;
+
+          // Shading function objects
+          for (size_t j = 0; j <= nStops; ++j)
+          {
+            startColour = stops.Item(j).GetColour();
+            stopColour = stops.Item(j+1).GetColour();
+            NewObj();
+            Out("<<");
+            Out("/FunctionType 2");
+            Out("/Domain [0 1]");
+            Out("/C0 [", false);
+            OutAscii(wxPdfColour(startColour).GetColourValue(), false);
+            Out("]");
+            Out("/C1 [", false);
+            OutAscii(wxPdfColour(stopColour).GetColourValue(), false);
+            Out("]");
+            Out("/N 1");
+            Out(">>");
+            Out("endobj");
+          }
+          if (nStops > 0)
+          {
+            // Stitching function object
+            NewObj();
+            Out("<<");
+            Out("/FunctionType 3");
+            Out("/Domain [0 1]");
+            Out("/Functions [ ", false);
+            for (int id = 0; id <= nStops ; ++id)
+            {
+              OutAscii(wxString::Format(wxS("%d 0 R "), idFuncShading+id), false);
+            }
+            Out("]");
+
+            Out("/Bounds [ ", false);
+            for (int id = 1; id <= nStops; ++id)
+            {
+              OutAscii(wxPdfUtility::Double2String(stops.Item(id).GetPosition(), 3), false);
+              Out(" ", false);
+            }
+            Out("]");
+
+            Out("/Encode [ ", false);
+            for (int id = 0; id <= nStops; ++id)
+            {
+              Out("0 1 ", false);
+            }
+            Out("]");
+            Out(">>");
+            Out("endobj");
+            idFuncShading = m_n;
+          }
+
+          // Shading object
+          NewObj();
+          Out("<<");
+          OutAscii(wxString::Format(wxS("/ShadingType %d"), ((patternStyle == wxPDF_PATTERNSTYLE_LINEAR_GRADIENT) ? 2 : 3)));
+          switch (wxPdfColour(startColour).GetColourType())
+          {
+            case wxPDF_COLOURTYPE_GRAY:
+              Out("/ColorSpace /DeviceGray");
+              break;
+            case wxPDF_COLOURTYPE_CMYK:
+              Out("/ColorSpace /DeviceCMYK");
+              break;
+            case wxPDF_COLOURTYPE_RGB:
+            default:
+              Out("/ColorSpace /DeviceRGB");
+              break;
+          }
+          if (patternStyle == wxPDF_PATTERNSTYLE_LINEAR_GRADIENT)
+          {
+            OutAscii(wxString(wxS("/Coords [")) +
+                     wxPdfUtility::Double2String(pattern->GetX1() * m_k, 3) + wxString(wxS(" ")) +
+                     wxPdfUtility::Double2String(pattern->GetY1() * m_k, 3) + wxString(wxS(" ")) +
+                     wxPdfUtility::Double2String(pattern->GetX2() * m_k, 3) + wxString(wxS(" ")) +
+                     wxPdfUtility::Double2String(pattern->GetY2() * m_k, 3) + wxString(wxS("]")));
+            OutAscii(wxString::Format(wxS("/Function %d 0 R"), idFuncShading));
+            Out("/Extend [true true] ");
+          }
+          else
+          {
+            OutAscii(wxString(wxS("/Coords [")) +
+              wxPdfUtility::Double2String(pattern->GetStartX() * m_k, 3) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(pattern->GetStartY() * m_k, 3) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(pattern->GetStartRadius() * m_k, 3) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(pattern->GetEndX() * m_k, 3) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(pattern->GetEndY() * m_k, 3) + wxString(wxS(" ")) +
+              wxPdfUtility::Double2String(pattern->GetEndRadius() * m_k, 3) + wxString(wxS("]")));
+            OutAscii(wxString::Format(wxS("/Function %d 0 R"), idFuncShading));
+            Out("/Extend [true true] ");
+          }
+          Out(">>");
+          Out("endobj");
+          int idShading = m_n;
+
+          // Pattern object
+          NewObj(pattern->GetObjIndex());
+          Out("<<");
+          Out("/Type /Pattern");
+          Out("/PatternType 2");
+          OutAscii(wxString::Format(wxS("/Shading %d 0 R"), idShading));
+
+          wxAffineMatrix2D matrix = pattern->GetMatrix();
+          wxAffineMatrix2D mirror;
+          if (m_yAxisOriginTop)
+          {
+            wxMatrix2D mat2D(1, 0, 0, -1);
+            wxPoint2DDouble tr(0, m_h*m_k);
+            mirror.Set(mat2D, tr);
+            if (!matrix.IsIdentity())
+            {
+              mirror.Concat(matrix);
+            }
+            matrix = mirror;
+          }
+
+          if (!matrix.IsIdentity())
+          {
+            wxMatrix2D mat2D;
+            wxPoint2DDouble tr;
+            matrix.Get(&mat2D, &tr);
+            OutAscii(wxString(wxS("/Matrix [")) +
+                     wxPdfUtility::Double2String(mat2D.m_11, 4) + wxString(" ") +
+                     wxPdfUtility::Double2String(mat2D.m_12, 4) + wxString(" ") +
+                     wxPdfUtility::Double2String(mat2D.m_21, 4) + wxString(" ") +
+                     wxPdfUtility::Double2String(mat2D.m_22, 4) + wxString(" ") +
+                     wxPdfUtility::Double2String(tr.m_x, 4) + wxString(" ") +
+                     wxPdfUtility::Double2String(tr.m_y, 4) + wxString("]"));
+          }
+          Out(">>");
+          Out("endobj");
+        }
+        break;
+
+      default:
+        if (patternStyle >= wxPDF_PATTERNSTYLE_FIRST_HATCH &&
+            patternStyle <= wxPDF_PATTERNSTYLE_LAST_HATCH)
+        {
+          NewObj(pattern->GetObjIndex());
+          Out("<<");
+          Out("/Type /Pattern");
+          Out("/PatternType 1");
+          Out("/PaintType 1");
+          Out("/TilingType 1");
+
+          OutAscii(wxString(wxS("/BBox [0 0 10 10]")));
+          OutAscii(wxString(wxS("/XStep 10")));
+          OutAscii(wxString(wxS("/YStep 10")));
+          OutAscii(wxString(wxS("/Resources << >>")));
+          wxString patternData;
+          double corrFactor = 1.0;
+          switch (pattern->GetPatternStyle())
+          {
+            case wxPDF_PATTERNSTYLE_BDIAGONAL_HATCH:
+              patternData = "0 0 m 10 10 l -1  9  m 1 11 l 9  -1  m 11 1 l";
+              break;
+            case wxPDF_PATTERNSTYLE_CROSSDIAG_HATCH:
+              patternData = "0 0 m 10 10 l 0 10 m 10 0 l";
+              break;
+            case wxPDF_PATTERNSTYLE_FDIAGONAL_HATCH:
+              patternData = "0 10 m 10 0 l -1 1 m 1 -1 l 9 11 m 11 9 l";
+              break;
+            case wxPDF_PATTERNSTYLE_CROSS_HATCH:
+              patternData = "0 5 m 10 5 l 5 0 m 5 10 l";
+              break;
+            case wxPDF_PATTERNSTYLE_HORIZONTAL_HATCH:
+              patternData = "0 5 m 10 5 l";
+              break;
+            case wxPDF_PATTERNSTYLE_VERTICAL_HATCH:
+              patternData = "5 0 m 5 10 l";
+              break;
+            case wxPDF_PATTERNSTYLE_HERRINGBONE_HATCH:
+              patternData = wxString(wxS("1.25 1.25 m 8.75 1.25 l 0 3.75 m 6.25 3.75 l 8.75 3.75 m 10 3.75 l")) +
+                wxS(" 0 6.25 m 3.75 6.25 l 6.25 6.25 m 10 6.25 l 0 8.75 m 1.25 8.75 l 3.75 8.75 m 10 8.75 l") +
+                wxS(" 1.25 0 m 1.25 3.75 l 1.25 6.25 m 1.25 10 l 3.75 0 m 3.75 1.25 l 3.75 3.75 m 3.75 10 l") +
+                wxS(" 6.25 1.25 m 6.25 8.75 l 8.75 0 m 8.75 6.25 l 8.75 8.75 m 8.75 10 l");
+              corrFactor = 4;
+              break;
+            case wxPDF_PATTERNSTYLE_BASKETWEAVE_HATCH:
+              patternData = wxString(wxS("0 1.25 m 10 1.25 l 0 6.25 m 10 6.25 l 3.75 0 m 3.75 10 l 8.75 0 m 8.75 10 l")) +
+                wxS(" 0 3.75 m 3.75 3.75 l 8.75 3.75 m 10 3.75 l 3.75 8.75 m 8.75 8.75 l") +
+                wxS(" 1.25 0 m 1.25 1.25 l 1.25 6.25 m 1.25 10 l 6.25 1.25 m 6.25 6.25 l");
+              corrFactor = 4;
+              break;
+            case wxPDF_PATTERNSTYLE_BRICK_HATCH:
+              patternData = "0 3 m 10 3 l 0 8 m 10 8 l 3 0 m 3 3 l 3 8 m 3 10 l 8 3 m 8 8 l";
+              corrFactor = 2;
+              break;
+
+            default:
+              break;
+          }
+          OutAscii(wxString(wxS("/Matrix [")) +
+            wxPdfUtility::Double2String((pattern->GetWidth() * m_k) / 10.0 * corrFactor, 4) + wxS(" 0 0 ") +
+            wxPdfUtility::Double2String((pattern->GetHeight() * m_k) / 10.0 * corrFactor, 4) + wxS(" 0 0]"));
+          wxString background;
+          if (pattern->HasFillColour())
+          {
+            background = wxPdfUtility::RGB2String(pattern->GetFillColour()) + wxS(" rg 0 0 10 10 re f ");
+          }
+          wxString sdata = wxString::Format(wxS("q ")) + background +
+            wxPdfUtility::RGB2String(pattern->GetDrawColour()) + wxS(" RG 2 J 0.5 w ") +
+            patternData + wxS(" S Q");
+          wxMemoryOutputStream mos;
+          mos.Write(sdata.ToAscii(), sdata.Length());
+          OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
+          Out(">>");
+          PutStream(mos);
+          Out("endobj");
+        }
+        break;
     }
-    else
-    {
-      OutAscii(wxString(wxS("/BBox [0 0 10 10]")));
-      OutAscii(wxString(wxS("/XStep 10")));
-      OutAscii(wxString(wxS("/YStep 10")));
-      OutAscii(wxString(wxS("/Resources << >>")));
-      wxString patternData;
-      double corrFactor = 1.0;
-      switch (pattern->GetPatternStyle())
-      {
-        case wxPDF_PATTERNSTYLE_BDIAGONAL_HATCH:
-          patternData = "0 0 m 10 10 l -1  9  m 1 11 l 9  -1  m 11 1 l";
-          break;
-        case wxPDF_PATTERNSTYLE_CROSSDIAG_HATCH:
-          patternData = "0 0 m 10 10 l 0 10 m 10 0 l";
-          break;
-        case wxPDF_PATTERNSTYLE_FDIAGONAL_HATCH:
-          patternData = "0 10 m 10 0 l -1 1 m 1 -1 l 9 11 m 11 9 l";
-          break;
-        case wxPDF_PATTERNSTYLE_CROSS_HATCH:
-          patternData = "0 5 m 10 5 l 5 0 m 5 10 l";
-          break;
-        case wxPDF_PATTERNSTYLE_HORIZONTAL_HATCH:
-          patternData = "0 5 m 10 5 l";
-          break;
-        case wxPDF_PATTERNSTYLE_VERTICAL_HATCH:
-          patternData = "5 0 m 5 10 l";
-          break;
-        case wxPDF_PATTERNSTYLE_HERRINGBONE_HATCH:
-          patternData = wxString(wxS("1.25 1.25 m 8.75 1.25 l 0 3.75 m 6.25 3.75 l 8.75 3.75 m 10 3.75 l")) +
-                        wxS(" 0 6.25 m 3.75 6.25 l 6.25 6.25 m 10 6.25 l 0 8.75 m 1.25 8.75 l 3.75 8.75 m 10 8.75 l") +
-                        wxS(" 1.25 0 m 1.25 3.75 l 1.25 6.25 m 1.25 10 l 3.75 0 m 3.75 1.25 l 3.75 3.75 m 3.75 10 l") +
-                        wxS(" 6.25 1.25 m 6.25 8.75 l 8.75 0 m 8.75 6.25 l 8.75 8.75 m 8.75 10 l");
-          corrFactor = 4;
-          break;
-        case wxPDF_PATTERNSTYLE_BASKETWEAVE_HATCH:
-          patternData = wxString(wxS("0 1.25 m 10 1.25 l 0 6.25 m 10 6.25 l 3.75 0 m 3.75 10 l 8.75 0 m 8.75 10 l")) +
-            wxS(" 0 3.75 m 3.75 3.75 l 8.75 3.75 m 10 3.75 l 3.75 8.75 m 8.75 8.75 l") +
-            wxS(" 1.25 0 m 1.25 1.25 l 1.25 6.25 m 1.25 10 l 6.25 1.25 m 6.25 6.25 l");
-          corrFactor = 4;
-          break;
-        case wxPDF_PATTERNSTYLE_BRICK_HATCH:
-          patternData = "0 3 m 10 3 l 0 8 m 10 8 l 3 0 m 3 3 l 3 8 m 3 10 l 8 3 m 8 8 l";
-          corrFactor = 2;
-          break;
-
-          // These pattern styles are not supported here, but still list them
-          // to avoid -Wswitch (and similar) warnings.
-        case wxPDF_PATTERNSTYLE_NONE:
-        case wxPDF_PATTERNSTYLE_IMAGE:
-        case wxPDF_PATTERNSTYLE_TEMPLATE:
-          break;
-      }
-      OutAscii(wxString(wxS("/Matrix [")) +
-        wxPdfUtility::Double2String((pattern->GetWidth() * m_k) / 10.0 * corrFactor, 4) + wxS(" 0 0 ") +
-        wxPdfUtility::Double2String((pattern->GetHeight() * m_k) / 10.0 * corrFactor, 4) + wxS(" 0 0]"));
-      wxString background;
-      if (pattern->HasFillColour())
-      { 
-        background = wxPdfUtility::RGB2String(pattern->GetFillColour()) + wxS(" rg 0 0 10 10 re f ");
-      }
-      wxString sdata = wxString::Format(wxS("q ")) + background +
-        wxPdfUtility::RGB2String(pattern->GetDrawColour()) + wxS(" RG 2 J 0.5 w ") +
-        patternData + wxS(" S Q");
-      wxMemoryOutputStream mos;
-      mos.Write(sdata.ToAscii(), sdata.Length());
-      OutAscii(wxString(wxS("/Length ")) + wxString::Format(wxS("%lu"), (unsigned long)CalculateStreamLength(mos.TellO())));
-      Out(">>");
-      PutStream(mos);
-    }
-    Out("endobj");
   }
 }